Refactor code for the persons logic
