+++ Data.rb
@@ -2,8 +2,14 @@ module Data
 
     @books = []
 
+    @persons = []
+
     def self.books
         @books
+    end
+
+    def self.persons
+        @persons
     end
 
 end
